Move interface from Util dialect to Flow dialect.
