+++ iree/compiler/Dialect/Flow/IR/FlowDialect.h
@@ -40,6 +40,9 @@ class FlowDialect : public Dialect {
   void registerTypes();
 };
 
+/// Register external models for PartitionableLoopsInterface.
+void registerPartitionableLoopsInterfaceModels(DialectRegistry &registry);
+
 }  // namespace Flow
 }  // namespace IREE
 }  // namespace iree_compiler
