+++ iree/compiler/Dialect/Flow/IR/PartitionableLoopsInterface.h
@@ -1,0 +1,17 @@
+// Copyright 2021 The IREE Authors
+//
+// Licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+#ifndef IREE_COMPILER_DIALECT_FLOW_IR_PARTITIONABLE_LOOPS_INTERFACE_H_
+#define IREE_COMPILER_DIALECT_FLOW_IR_PARTITIONABLE_LOOPS_INTERFACE_H_
+
+#include "iree/compiler/Dialect/Flow/IR/FlowDialect.h"
+#include "llvm/ADT/SmallVector.h"
+#include "mlir/IR/OpDefinition.h"
+#include "mlir/Support/LLVM.h"
+
+#include "iree/compiler/Dialect/Flow/IR/PartitionableLoopsInterface.h.inc"  // IWYU pragma: export
+
+#endif  // IREE_COMPILER_DIALECT_FLOW_IR_PARTITIONABLE_LOOPS_INTERFACE_H_
