+++ iree/compiler/Dialect/Flow/Transforms/CMakeLists.txt
@@ -50,6 +50,7 @@ iree_cc_library(
     "PromoteI1ToI8Pass.cpp"
     "StripAndSplatConstantVariables.cpp"
     "StripSignednessPass.cpp"
+    "TestPartitionableLoopsInterface.cpp"
     "TypeConverter.cpp"
     "VerifyInputLegality.cpp"
   DEPS
