+++ iree/compiler/Dialect/Flow/Transforms/TestPartitionableLoopsInterface.cpp
@@ -4,9 +4,9 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/compiler/Dialect/Util/IR/PartitionableLoopsInterface.h"
+#include "iree/compiler/Dialect/Flow/IR/PartitionableLoopsInterface.h"
+#include "iree/compiler/Dialect/Flow/Transforms/Passes.h"
 #include "iree/compiler/Dialect/Util/IR/UtilOps.h"
-#include "iree/compiler/Dialect/Util/Transforms/Passes.h"
 #include "mlir/Dialect/StandardOps/IR/Ops.h"
 #include "mlir/Transforms/GreedyPatternRewriteDriver.h"
 
@@ -15,7 +15,7 @@ static const char kAttributeName[] = "__test_interface__";
 namespace mlir {
 namespace iree_compiler {
 namespace IREE {
-namespace Util {
+namespace Flow {
 
 namespace {
 
@@ -43,7 +43,8 @@ struct TestPartitionableLoopsInterfacePattern
     }
     auto type = RankedTensorType::get(numLoops, rewriter.getIndexType());
     auto constantAttr = DenseIntElementsAttr::get(type, loopInfo);
-    rewriter.create<UnfoldableConstantOp>(interfaceOp.getLoc(), constantAttr);
+    rewriter.create<Util::UnfoldableConstantOp>(interfaceOp.getLoc(),
+                                                constantAttr);
     rewriter.updateRootInPlace(
         interfaceOp, [&] { interfaceOp->removeAttr(kAttributeName); });
     return success();
@@ -54,7 +55,7 @@ struct TestPartitionableLoopsInterfacePass
     : public PassWrapper<TestPartitionableLoopsInterfacePass,
                          OperationPass<void>> {
   StringRef getArgument() const override {
-    return "iree-util-test-partitionable-loops-interface";
+    return "iree-flow-test-partitionable-loops-interface";
   }
 
   StringRef getDescription() const override {
@@ -63,7 +64,7 @@ struct TestPartitionableLoopsInterfacePass
   }
 
   void getDependentDialects(DialectRegistry &registry) const override {
-    registry.insert<UtilDialect>();
+    registry.insert<FlowDialect>();
   }
 
   void runOnOperation() override {
@@ -85,8 +86,7 @@ createTestPartitionableLoopsInterfacePass() {
 
 static PassRegistration<TestPartitionableLoopsInterfacePass> pass;
 
-}  // namespace Util
-
+}  // namespace Flow
 }  // namespace IREE
 }  // namespace iree_compiler
 }  // namespace mlir
