+++ iree/compiler/Dialect/Flow/Transforms/test/CMakeLists.txt
@@ -35,6 +35,7 @@ iree_lit_test_suite(
     "promote_i1_to_i8.mlir"
     "strip_and_splat_constant_variables.mlir"
     "strip_signedness.mlir"
+    "test_partitionable_loops_interface.mlir"
     "transformation.mlir"
     "verify_input_ir.mlir"
   DATA
