+++ iree/compiler/Dialect/Flow/Transforms/test/test_partitionable_loops_interface.mlir
@@ -1,4 +1,4 @@
-// RUN: iree-opt -iree-util-test-partitionable-loops-interface -split-input-file %s | IreeFileCheck %s
+// RUN: iree-opt -iree-flow-test-partitionable-loops-interface -split-input-file %s | IreeFileCheck %s
 
 #map1 = affine_map<(d0, d1, d2) -> (d0, d1, d2)>
 #map2 = affine_map<(d0, d1, d2) -> (d0, d2)>
