+++ iree/compiler/Dialect/Util/IR/BUILD
@@ -20,7 +20,6 @@ td_library(
     name = "td_files",
     srcs = enforce_glob(
         [
-            "PartitionableLoopsInterface.td",
             "UtilAttrs.td",
             "UtilBase.td",
             "UtilInterfaces.td",
@@ -42,7 +41,6 @@ cc_library(
     name = "IR",
     srcs = [
         "ClosureOpUtils.cpp",
-        "PartitionableLoopsInterface.cpp",
         "UtilAttrs.cpp",
         "UtilDialect.cpp",
         "UtilOpFolders.cpp",
@@ -52,7 +50,6 @@ cc_library(
     ],
     hdrs = [
         "ClosureOpUtils.h",
-        "PartitionableLoopsInterface.h",
         "UtilDialect.h",
         "UtilOps.h",
         "UtilOps.h.inc",
@@ -60,8 +57,6 @@ cc_library(
         "UtilTypes.h",
     ],
     textual_hdrs = [
-        "PartitionableLoopsInterface.cpp.inc",
-        "PartitionableLoopsInterface.h.inc",
         "UtilAttrs.cpp.inc",
         "UtilAttrs.h.inc",
         "UtilAttrInterfaces.cpp.inc",
@@ -72,16 +67,13 @@ cc_library(
         "UtilTypeInterfaces.h.inc",
     ],
     deps = [
-        ":PartitionableLoopsInterfaceGen",
         ":UtilAttrsGen",
         ":UtilInterfacesGen",
         ":UtilOpsGen",
-        "//llvm-external-projects/iree-dialects:IREELinalgExtDialect",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:ArithmeticDialect",
         "@llvm-project//mlir:ControlFlowInterfaces",
         "@llvm-project//mlir:IR",
-        "@llvm-project//mlir:LinalgInterfaces",
         "@llvm-project//mlir:MemRefDialect",
         "@llvm-project//mlir:Parser",
         "@llvm-project//mlir:SideEffectInterfaces",
@@ -91,23 +83,6 @@ cc_library(
         "@llvm-project//mlir:TensorDialect",
         "@llvm-project//mlir:Transforms",
     ],
-)
-
-gentbl_cc_library(
-    name = "PartitionableLoopsInterfaceGen",
-    tbl_outs = [
-        (
-            ["-gen-op-interface-decls"],
-            "PartitionableLoopsInterface.h.inc",
-        ),
-        (
-            ["-gen-op-interface-defs"],
-            "PartitionableLoopsInterface.cpp.inc",
-        ),
-    ],
-    tblgen = "@llvm-project//mlir::mlir-tblgen",
-    td_file = "PartitionableLoopsInterface.td",
-    deps = ["@llvm-project//mlir:OpBaseTdFiles"],
 )
 
 gentbl_cc_library(
