+++ iree/compiler/Dialect/Util/IR/UtilDialect.h
@@ -38,9 +38,6 @@ class UtilDialect : public Dialect {
 
 void registerUtilExternalModels(DialectRegistry& registry);
 
-/// Register external models for PartitionableLoopsInterface.
-void registerPartitionableLoopsInterfaceModels(DialectRegistry& registry);
-
 }  // namespace Util
 }  // namespace IREE
 }  // namespace iree_compiler
