+++ iree/compiler/Dialect/Util/Transforms/BUILD
@@ -23,7 +23,6 @@ cc_library(
         "Patterns.cpp",
         "SimplifyGlobalAccesses.cpp",
         "TestFloatRangeAnalysis.cpp",
-        "TestPartitionableLoopsInterface.cpp",
     ],
     hdrs = [
         "Passes.h",
