+++ iree/compiler/Dialect/Util/Transforms/Passes.h
@@ -27,8 +27,6 @@ std::unique_ptr<OperationPass<void>> createFixedPointIteratorPass(
 
 // Test passes.
 std::unique_ptr<OperationPass<void>> createTestFloatRangeAnalysis();
-std::unique_ptr<OperationPass<void>>
-createTestPartitionableLoopsInterfacePass();
 
 // Register all Passes
 // TODO: Switch this directory to declarative registration.
@@ -42,7 +40,6 @@ inline void registerTransformPasses() {
   createSimplifyGlobalAccessesPass();
   createFixedPointIteratorPass(OpPassManager("dummy_op"));
   createTestFloatRangeAnalysis();
-  createTestPartitionableLoopsInterfacePass();
 }
 
 }  // namespace Util
