+++ iree/compiler/Dialect/Util/Transforms/test/BUILD
@@ -27,7 +27,6 @@ iree_lit_test_suite(
             "simplify_global_accesses.mlir",
             "test_float_range_analysis.mlir",
             "test_float_range_analysis_linalg.mlir",
-            "test_partitionable_loops_interface.mlir",
         ],
         include = ["*.mlir"],
     ),
