+++ iree/tools/init_iree_dialects.h
@@ -46,9 +46,9 @@ inline void registerIreeDialects(DialectRegistry &registry) {
                   IREE::PYDM::IREEPyDMDialect>();
   // clang-format on
 
+  IREE::Flow::registerPartitionableLoopsInterfaceModels(registry);
   IREE::LinalgExt::registerTiledOpInterfaceExternalModels(registry);
   IREE::Util::registerUtilExternalModels(registry);
-  IREE::Util::registerPartitionableLoopsInterfaceModels(registry);
   registerCodegenInterfaces(registry);
 }
